+++ duckscript_sdk/src/sdk/std/string/mod.rs
@@ -4,6 +4,7 @@ mod equals;
 mod indexof;
 mod is_empty;
 mod last_indexof;
+mod length;
 mod starts_with;
 mod substring;
 mod trim;
@@ -25,6 +26,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(indexof::create(&package))?;
     commands.set(is_empty::create(&package))?;
     commands.set(last_indexof::create(&package))?;
+    commands.set(length::create(&package))?;
     commands.set(starts_with::create(&package))?;
     commands.set(substring::create(&package))?;
     commands.set(trim::create(&package))?;